+++ R/geom_milb_logos.R
@@ -1,21 +1,18 @@
-#' @name geom_cap_logos
+#' @name geom_milb_logos
 #'
-#' @title
-#' ggplot2 Layer for Visualizing MLB Team Cap Logos
+#' @title ggplot2 Layer for Visualizing MiLB Team Logos
 #'
-#' @description `geom_mlb_dark_cap_logos()` and `geom_mlb_light_cap_logos()`
-#'   are used to plot MLB team cap and league logos instead of points in a
-#'   ggplot. It requires x, y aesthetics as well as a valid MLB team
-#'   abbreviation. The latter can be checked with [`valid_team_names()`] but is
-#'   also cleaned before being plotted.
+#' @description `geom_milb_logos()`, `geom_milb_light_cap_logos()`, `geom_milb_dot_logos()` are used to
+#'   plot MiLB team instead of points in a ggplot. It requires
+#'   x, y aesthetics as well as a valid MiLB team name
 #'
 #' @inheritParams ggplot2::geom_point
 #' @section Aesthetics:
-#' `geom_mlb_dark_cap_logos()` and `geom_mlb_light_cap_logos()` understand the following aesthetics:
+#' `geom_milb_logos()`, `geom_milb_light_cap_logos()`, `geom_milb_dot_logos()` understand the following aesthetics:
 #' \describe{
 #'   \item{`x`}{ - The x-coordinate. Required.}
 #'   \item{`y`}{ - The y-coordinate. Required.}
-#'   \item{`team_abbr`}{ - The team abbreviation. Need to use Savant's abbreviation. Required.}
+#'   \item{`team_name`}{ - The team name. Need to use the full team name. Required.}
 #'   \item{`alpha = NULL`}{ - The alpha channel, i.e. transparency level, as a numerical value between 0 and 1.}
 #'   \item{`colour = NULL`}{ - The image will be colourized with this colour. Use the special character `"b/w"` to set it to black and white. For more information on valid colour names in ggplot2 see <https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill>}
 #'   \item{`angle = 0`}{ - The angle of the image as a numerical value between 0° and 360°.}
@@ -45,42 +42,35 @@
 #' library(mlbplotR)
 #' library(ggplot2)
 #'
-#' team_abbr <- valid_team_names()
-#' # remove conference logos from this example
-#' team_abbr <- team_abbr[!team_abbr %in% c("NL", "AL", "MLB")]
+#' team_names <- c("Kannapolis Cannon Ballers", "Charlotte Knights",
+#'                 "Chesapeake Baysox", "Durham Bulls", "Montgomery Biscuits", "Las Vegas Aviators",
+#'                 "Lehigh Valley IronPigs", "Richmond Flying Squirrels", "Round Rock Express",
+#'                 "Frisco RoughRiders", "Hub City Spartanburgers", "Hickory Crawdads")
 #'
 #' df <- data.frame(
-#'   a = rep(1:6, 5),
-#'   b = sort(rep(1:5, 6), decreasing = TRUE),
-#'   teams = team_abbr
+#'   a = rep(1:4, 3),
+#'   b = sort(rep(1:3, 4), decreasing = TRUE),
+#'   teams = team_names
 #' )
 #'
 #' # keep alpha == 1 for all teams including an "A"
-#' matches <- grepl("A", team_abbr)
+#' matches <- grepl("A|a", team_names)
 #' df$alpha <- ifelse(matches, 1, 0.2)
 #' # also set a custom fill colour for the non "A" teams
 #' df$colour <- ifelse(matches, NA, "gray")
 #'
 #' # scatterplot of all logos
 #' ggplot(df, aes(x = a, y = b)) +
-#'   geom_mlb_dark_cap_logos(aes(team_abbr = teams), height = 0.075) +
+#'   geom_milb_logos(aes(team_name = teams), height = 0.1) +
 #'   geom_label(aes(label = teams), nudge_y = -0.35, alpha = 0.5) +
 #'   theme_void()
 #'
-#' # apply alpha via an aesthetic from inside the dataset `df`
-#' # please note that you have to add scale_alpha_identity() to use the alpha
-#' # values in your dataset!
-#' ggplot(df, aes(x = a, y = b)) +
-#'   geom_mlb_dark_cap_logos(aes(team_abbr = teams, alpha = alpha), height = 0.075) +
-#'   geom_label(aes(label = teams), nudge_y = -0.35, alpha = 0.5) +
-#'   scale_alpha_identity() +
-#'   theme_void()
 #'
 #' # apply alpha and colour via an aesthetic from inside the dataset `df`
 #' # please note that you have to add scale_alpha_identity() as well as
 #' # scale_colour_identity() to use the alpha and colour values in your dataset!
 #' ggplot(df, aes(x = a, y = b)) +
-#'   geom_mlb_light_cap_logos(aes(team_abbr = teams, alpha = alpha, colour = colour), height = 0.075) +
+#'   geom_milb_light_cap_logos(aes(team_name = teams, alpha = alpha, colour = colour), height = 0.1) +
 #'   geom_label(aes(label = teams), nudge_y = -0.35, alpha = 0.5) +
 #'   scale_alpha_identity() +
 #'   scale_colour_identity() +
@@ -88,23 +78,84 @@
 #'
 #' # apply alpha as constant for all logos
 #' ggplot(df, aes(x = a, y = b)) +
-#'   geom_mlb_light_cap_logos(aes(team_abbr = teams), height = 0.075, alpha = 0.6) +
+#'   geom_milb_dot_logos(aes(team_name = teams), height = 0.15, alpha = 0.6) +
 #'   geom_label(aes(label = teams), nudge_y = -0.35, alpha = 0.5) +
 #'   theme_void()
 #'
-#' # it's also possible to plot league logos
-#' league <- data.frame(a = 1:3, b = 0, teams = c("AL", "NL", "MLB"))
-#' ggplot(league, aes(x = a, y = b)) +
-#'   geom_mlb_dark_cap_logos(aes(team_abbr = teams), width = 0.3) +
-#'   geom_label(aes(label = teams), nudge_y = -0.4, alpha = 0.5) +
-#'   coord_cartesian(xlim = c(0.5,3.5), ylim = c(-0.75,.75)) +
-#'   theme_void()
 #' }
 NULL
 
-#' @rdname geom_cap_logos
+#' @rdname geom_milb_logos
 #' @export
-geom_mlb_dark_cap_logos <- function(mapping = NULL, data = NULL,
+geom_milb_logos <- function(mapping = NULL, data = NULL,
+                               stat = "identity", position = "identity",
+                               ...,
+                               nudge_x = 0,
+                               nudge_y = 0,
+                               na.rm = FALSE,
+                               show.legend = FALSE,
+                               inherit.aes = TRUE) {
+
+  if (!missing(nudge_x) || !missing(nudge_y)) {
+    if (!missing(position)) {
+      cli::cli_abort(c(
+        "both {.arg position} and {.arg nudge_x}/{.arg nudge_y} are supplied",
+        "i" = "Only use one approach to alter the position"
+      ))
+    }
+
+    position <- ggplot2::position_nudge(nudge_x, nudge_y)
+  }
+
+  ggplot2::layer(
+    data = data,
+    mapping = mapping,
+    stat = stat,
+    geom = GeomMiLBlogo,
+    position = position,
+    show.legend = show.legend,
+    inherit.aes = inherit.aes,
+    params = list(
+      na.rm = na.rm,
+      ...
+    )
+  )
+}
+
+#' @rdname mlbplotR-package
+#' @export
+GeomMiLBlogo <- ggplot2::ggproto(
+  "GeomMiLBlogo", ggplot2::Geom,
+  required_aes = c("x", "y", "team_name"),
+  # non_missing_aes = c(""),
+  default_aes = ggplot2::aes(
+    alpha = NULL, colour = NULL, angle = 0, hjust = 0.5,
+    vjust = 0.5, width = 1.0, height = 1.0
+  ),
+  draw_panel = function(data, panel_params, coord, na.rm = FALSE) {
+    data <- coord$transform(data, panel_params)
+
+    grobs <- lapply(seq_along(data$team_name), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "milb_logo")
+
+    class(grobs) <- "gList"
+
+    grid::gTree(children = grobs)
+  },
+  draw_key = function(...) grid::nullGrob()
+)
+
+
+
+
+
+
+
+
+
+
+#' @rdname geom_milb_logos
+#' @export
+geom_milb_light_cap_logos <- function(mapping = NULL, data = NULL,
                            stat = "identity", position = "identity",
                            ...,
                            nudge_x = 0,
@@ -128,7 +179,7 @@ geom_mlb_dark_cap_logos <- function(mapping = NULL, data = NULL,
     data = data,
     mapping = mapping,
     stat = stat,
-    geom = GeomMLBdarkcaplogo,
+    geom = GeomMiLBlightcaplogo,
     position = position,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
@@ -141,9 +192,9 @@ geom_mlb_dark_cap_logos <- function(mapping = NULL, data = NULL,
 
 #' @rdname mlbplotR-package
 #' @export
-GeomMLBdarkcaplogo <- ggplot2::ggproto(
-  "GeomMLBdarkcaplogo", ggplot2::Geom,
-  required_aes = c("x", "y", "team_abbr"),
+GeomMiLBlightcaplogo <- ggplot2::ggproto(
+  "GeomMiLBlightcaplogo", ggplot2::Geom,
+  required_aes = c("x", "y", "team_name"),
   # non_missing_aes = c(""),
   default_aes = ggplot2::aes(
     alpha = NULL, colour = NULL, angle = 0, hjust = 0.5,
@@ -152,9 +203,7 @@ GeomMLBdarkcaplogo <- ggplot2::ggproto(
   draw_panel = function(data, panel_params, coord, na.rm = FALSE) {
     data <- coord$transform(data, panel_params)
 
-    data$team_abbr <- clean_team_abbrs(as.character(data$team_abbr), keep_non_matches = FALSE)
-
-    grobs <- lapply(seq_along(data$team_abbr), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "dark_cap")
+    grobs <- lapply(seq_along(data$team_name), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "milb_light_cap")
 
     class(grobs) <- "gList"
 
@@ -166,16 +215,18 @@ GeomMLBdarkcaplogo <- ggplot2::ggproto(
 
 
 
-#' @rdname geom_cap_logos
+
+
+#' @rdname geom_milb_logos
 #' @export
-geom_mlb_light_cap_logos <- function(mapping = NULL, data = NULL,
-                                    stat = "identity", position = "identity",
-                                    ...,
-                                    nudge_x = 0,
-                                    nudge_y = 0,
-                                    na.rm = FALSE,
-                                    show.legend = FALSE,
-                                    inherit.aes = TRUE) {
+geom_milb_dot_logos <- function(mapping = NULL, data = NULL,
+                                stat = "identity", position = "identity",
+                                ...,
+                                nudge_x = 0,
+                                nudge_y = 0,
+                                na.rm = FALSE,
+                                show.legend = FALSE,
+                                inherit.aes = TRUE) {
 
   if (!missing(nudge_x) || !missing(nudge_y)) {
     if (!missing(position)) {
@@ -192,7 +243,7 @@ geom_mlb_light_cap_logos <- function(mapping = NULL, data = NULL,
     data = data,
     mapping = mapping,
     stat = stat,
-    geom = GeomMLBlightcaplogo,
+    geom = GeomMiLBdotlogo,
     position = position,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
@@ -205,9 +256,9 @@ geom_mlb_light_cap_logos <- function(mapping = NULL, data = NULL,
 
 #' @rdname mlbplotR-package
 #' @export
-GeomMLBlightcaplogo <- ggplot2::ggproto(
-  "GeomMLBlightcaplogo", ggplot2::Geom,
-  required_aes = c("x", "y", "team_abbr"),
+GeomMiLBdotlogo <- ggplot2::ggproto(
+  "GeomMiLBdotlogo", ggplot2::Geom,
+  required_aes = c("x", "y", "team_name"),
   # non_missing_aes = c(""),
   default_aes = ggplot2::aes(
     alpha = NULL, colour = NULL, angle = 0, hjust = 0.5,
@@ -216,9 +267,7 @@ GeomMLBlightcaplogo <- ggplot2::ggproto(
   draw_panel = function(data, panel_params, coord, na.rm = FALSE) {
     data <- coord$transform(data, panel_params)
 
-    data$team_abbr <- clean_team_abbrs(as.character(data$team_abbr), keep_non_matches = FALSE)
-
-    grobs <- lapply(seq_along(data$team_abbr), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "light_cap")
+    grobs <- lapply(seq_along(data$team_name), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "milb_dot")
 
     class(grobs) <- "gList"