+++ R/gt_helpers.R
@@ -11,7 +11,15 @@
 #'
 #' @param gt_object An existing gt table object of class `gt_tbl`
 #' @param columns The columns wherein changes to cell data colors should occur.
+#'   Argument has no effect if `locations` is not `NULL`.
 #' @param height The absolute height (px) of the image in the table cell
+#' @param locations If `NULL` (the default), the function will render
+#'   logos in argument `columns`.
+#'   Otherwise, the cell or set of cells to be associated with the team name
+#'   transformation. Only the [gt::cells_body()], [gt::cells_stub()],
+#'   [gt::cells_column_labels()], and [gt::cells_row_groups()] helper functions
+#'   can be used here. We can enclose several of these calls within a `list()`
+#'   if we wish to make the transformation happen at different locations.
 #' @return An object of class `gt_tbl`.
 #' @export
 #' @examples
@@ -29,14 +37,15 @@
 #'  gt_fmt_mlb_scoreboard_logo(columns = "scoreboard_logo") %>%
 #'  gt_fmt_mlb_dot_logo(columns = "dot_logo")
 
-gt_fmt_mlb_logo <- function(gt_object, columns, height = 30){
+gt_fmt_mlb_logo <- function(gt_object, columns, height = 30, locations = NULL){
 
   stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
 
   gt_mlbplotR_image(
     gt_object = gt_object,
-    columns = columns,
+    columns = {{ columns }},
     height = height,
+    locations = locations,
     type = "mlb_logo"
   )
 
@@ -45,14 +54,15 @@ gt_fmt_mlb_logo <- function(gt_object, columns, height = 30){
 #' @rdname gt_mlb
 #' @export
 
-gt_fmt_mlb_scoreboard_logo <- function(gt_object, columns, height = 30){
+gt_fmt_mlb_scoreboard_logo <- function(gt_object, columns, height = 30, locations = NULL){
 
   stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
 
   gt_mlbplotR_image(
     gt_object = gt_object,
-    columns = columns,
+    columns = {{ columns }},
     height = height,
+    locations = locations,
     type = "scoreboard_logo"
   )
 }
@@ -61,14 +71,15 @@ gt_fmt_mlb_scoreboard_logo <- function(gt_object, columns, height = 30){
 #' @rdname gt_mlb
 #' @export
 
-gt_fmt_mlb_dot_logo <- function(gt_object, columns, height = 30){
+gt_fmt_mlb_dot_logo <- function(gt_object, columns, height = 30, locations = NULL){
 
   stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
 
   gt_mlbplotR_image(
     gt_object = gt_object,
-    columns = columns,
+    columns = {{ columns }},
     height = height,
+    locations = locations,
     type = "dot_logo"
   )
 }
@@ -78,29 +89,64 @@ gt_fmt_mlb_dot_logo <- function(gt_object, columns, height = 30){
 gt_mlbplotR_image <- function(gt_object,
                               columns,
                               height = 30,
+                              locations = NULL,
                               type = c("mlb_logo", "scoreboard_logo", "dot_logo")){
 
   rlang::check_installed("gt (>= 0.8.0)", "to render images in gt tables.")
 
   type <- match.arg(type)
 
-  locations <- gt::cells_body({{ columns }})
+  if(is.null(locations)){
+    locations <- gt::cells_body({{ columns }})
+  }
 
   if (is.numeric(height)) {
     height <- paste0(height, "px")
   }
 
-  gt::text_transform(
-    data = gt_object,
-    locations = locations,
-    fn = function(x){
-      team_abbr <- clean_team_abbrs(as.character(x), keep_non_matches = FALSE)
-      # Create the image URI
-      uri <- get_image_uri(team_abbr = team_abbr, type = type)
-      # Generate the Base64-encoded image and place it within <img> tags
-      paste0("<img src=\"", uri, "\" style=\"height:", height, ";\">")
-    }
-  )
+  if (type %in% c("mlb_logo", "scoreboard_logo")) {
+    gt::text_transform(
+      data = gt_object,
+      locations = locations,
+      fn = function(x){
+        team_abbr <- clean_team_abbrs(as.character(x), keep_non_matches = FALSE)
+        # Create the image URI
+        uri <- get_image_uri(team_abbr = team_abbr, type = type)
+        # Generate the Base64-encoded image and place it within <img> tags
+        out <- paste0("<img src=\"", uri, "\" style=\"height:", height, ";\">")
+        out <- lapply(out, gt::html)
+        # If the image uri returns NA we didn't find a match. We will return the
+        # actual value then to allow the user to call gt::sub_missing()
+        out[is.na(uri)] <- x[is.na(uri)]
+        out
+      }
+    )
+  } else if (type == "dot_logo") {
+
+    gt::text_transform(
+      data = gt_object,
+      locations = locations,
+      fn = function(x){
+        image_urls <- vapply(
+          x,
+          FUN.VALUE = character(1),
+          USE.NAMES = FALSE,
+          FUN = function(team_abbr) {
+
+            if (team_abbr %in% names(dot_logo_list)) {
+              dot_logo_list[[team_abbr]]
+            } else {
+              # Return the MLB logo if there isn't a match
+              dot_logo_list[["MLB"]]
+            }
+
+          }
+        )
+        gt::web_image(image_urls, height = height)
+      }
+    )
+
+  }
 
 }
 
@@ -121,6 +167,8 @@ get_image_uri <- function(team_abbr, type = c("mlb_logo", "scoreboard_logo", "do
     FUN.VALUE = character(1),
     USE.NAMES = FALSE,
     FUN = function(team) {
+      # every non match will return NULL which is when we want NA
+      if (is.null(lookup_list[[team]])) return(NA_character_)
       paste0(
         "data:", "image/png",
         ";base64,", base64enc::base64encode(lookup_list[[team]])
@@ -130,13 +178,121 @@ get_image_uri <- function(team_abbr, type = c("mlb_logo", "scoreboard_logo", "do
 }
 
 
+#' @name gt_milb
+#' @title
+#' Add MiLB team logos into rows of a `gt` table
+#' @description
+#' The `gt_fmt_milb_logo` and `gt_fmt_milb_dot_logo` functions take an existing
+#' `gt_tbl` object and converts MiLB team names into team logos.
+#' This is a wrapper around
+#' [`gtExtras::gt_image_rows()`](https://jthomasmock.github.io/gtExtras/reference/gt_img_rows.html)
+#' written by Tom Mock, which is a wrapper around `gt::text_transform()` + `gt::web_image()`/
+#' `gt::local_image()` with the necessary boilerplate already applied.
+#'
+#' @param gt_object An existing gt table object of class `gt_tbl`
+#' @param columns The columns wherein changes to cell data colors should occur.
+#'   Argument has no effect if `locations` is not `NULL`.
+#' @param height The absolute height (px) of the image in the table cell
+#' @param locations If `NULL` (the default), the function will render
+#'   logos in argument `columns`.
+#'   Otherwise, the cell or set of cells to be associated with the team name
+#'   transformation. Only the [gt::cells_body()], [gt::cells_stub()],
+#'   [gt::cells_column_labels()], and [gt::cells_row_groups()] helper functions
+#'   can be used here. We can enclose several of these calls within a `list()`
+#'   if we wish to make the transformation happen at different locations.
+#' @return An object of class `gt_tbl`.
+#' @examples
+#' \donttest{
+#' library(gt)
+#' library(mlbplotR)
+#' gt_milb_example <- mlbplotR::load_milb_teams() %>%
+#'   dplyr::filter(parent_org_name == "Texas Rangers") %>%
+#'   dplyr::mutate(dot = team_name) %>%
+#'   dplyr::select(team_name, dot, team_location, team_mascot) %>%
+#'   gt::gt() %>%
+#'   gt_fmt_milb_logo(columns = "team_name") %>%
+#'   gt_fmt_milb_dot_logo(columns = "dot")
+#' }
+#' @export
+gt_fmt_milb_logo <- function(gt_object,
+                             columns,
+                             height = 30,
+                             locations = NULL) {
+
+  stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
+
+
+  rlang::check_installed("gt (>= 0.8.0)", "to render images in gt tables.")
+
+  if(is.null(locations)){
+    locations <- gt::cells_body({{ columns }})
+  }
+
+  gt::text_transform(
+    data = gt_object,
+    locations = locations,
+    fn = function(mlb_id){
+      milb_logo_map <- load_milb_teams()
+      image_urls <- vapply(
+        mlb_id,
+        FUN.VALUE = character(1),
+        USE.NAMES = FALSE,
+        FUN = function(team_name_full) {
+          ret <- milb_logo_map$team_logo[milb_logo_map$team_name == team_name_full]
+          if(length(ret) == 0 || is.na(ret)) ret <- NA_character_
+          ret
+        }
+      )
+      gt::web_image(image_urls, height = height)
+    }
+  )
+
+}
+
+#' @rdname gt_milb
+#' @export
+gt_fmt_milb_dot_logo <- function(gt_object,
+                                 columns,
+                                 height = 30,
+                                 locations = NULL) {
+
+  stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
+
+
+  rlang::check_installed("gt (>= 0.8.0)", "to render images in gt tables.")
+
+  if(is.null(locations)){
+    locations <- gt::cells_body({{ columns }})
+  }
+
+  gt::text_transform(
+    data = gt_object,
+    locations = locations,
+    fn = function(mlb_id){
+      milb_logo_map <- load_milb_teams()
+      image_urls <- vapply(
+        mlb_id,
+        FUN.VALUE = character(1),
+        USE.NAMES = FALSE,
+        FUN = function(team_name_full) {
+          ret <- milb_logo_map$team_dot_logo[milb_logo_map$team_name == team_name_full]
+          if(length(ret) == 0 || is.na(ret)) ret <- NA_character_
+          ret
+        }
+      )
+      gt::web_image(image_urls, height = height)
+    }
+  )
+
+}
 
 
 
-#' Render Player Headshots in 'gt' Tables
+#' @name gt_mlb_headshots
+#' @title Render Player Headshots in 'gt' Tables
 #'
 #' @description
-#' `gt_fmt_mlb_headshot` takes an existing `gt_tbl` object and converts player ids into headshots.
+#' `gt_fmt_mlb_headshot`, `gt_fmt_mlb_dot_headshot`, and `gt_fmt_milb_dot_headshot` take an existing `gt_tbl` object and converts player ids into headshots.
 #' This is a wrapper around
 #' [`gtExtras::gt_image_rows()`](https://jthomasmock.github.io/gtExtras/reference/gt_img_rows.html)
 #' written by Tom Mock, which is a wrapper around `gt::text_transform()` + `gt::web_image()`/
@@ -146,45 +302,135 @@ get_image_uri <- function(team_abbr, type = c("mlb_logo", "scoreboard_logo", "do
 #' @param columns The columns wherein changes to cell data colors should occur.
 #'   Has no effect if `locations` is not `NULL`
 #' @param height The absolute height (px) of the image in the table cell
-#' @param na_headshot_to_logo should NA/non matches return the MLB logo instead
-#'   of a grayed out blank headshot? Defaults to `TRUE`
+#' @param locations If `NULL` (the default), the function will render
+#'   logos in argument `columns`.
+#'   Otherwise, the cell or set of cells to be associated with the team name
+#'   transformation. Only the [gt::cells_body()], [gt::cells_stub()],
+#'   [gt::cells_column_labels()], and [gt::cells_row_groups()] helper functions
+#'   can be used here. We can enclose several of these calls within a `list()`
+#'   if we wish to make the transformation happen at different locations.
 #'
 #' @return An object of class `gt_tbl`.
-#' @export
 #' @examples
+#' \donttest{
 #' library(gt)
 #' library(mlbplotR)
-#' gt_headshot_example <- mlbplotR::load_headshots() %>%
-#'   head(5) %>%
-#'   dplyr::select(player_name, savant_id) %>%
+#' gt_headshot_example <- data.frame(
+#'   player_name = c(
+#'     "Nathan Eovaldi", "Marcus Semien", "Corey Seager", "Jacob deGrom", "Chris Martin"
+#'   ),
+#'   savant_id1 = c(543135, 543760, 608369, 594798, 455119),
+#'   savant_id2 = c(543135, 543760, 608369, 594798, 455119)
+#'   ) %>%
 #'   gt::gt() %>%
-#'   gt_fmt_mlb_headshot(columns = "savant_id")
+#'   gt_fmt_mlb_headshot(columns = "savant_id1") %>%
+#'   gt_fmt_mlb_dot_headshot(columns = "savant_id2")
+#' }
+NULL
+
+#' @rdname gt_mlb_headshots
 #' @export
 gt_fmt_mlb_headshot <- function(gt_object,
                                 columns,
                                 height = 30,
-                                na_headshot_to_logo = TRUE) {
+                                locations = NULL) {
 
   stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
 
 
   rlang::check_installed("gt (>= 0.8.0)", "to render images in gt tables.")
 
-  locations <- gt::cells_body({{ columns }})
+  if(is.null(locations)){
+    locations <- gt::cells_body({{ columns }})
+  }
 
   gt::text_transform(
     data = gt_object,
     locations = locations,
     fn = function(mlb_id){
-      headshot_map <- load_headshots()
       image_urls <- vapply(
         mlb_id,
         FUN.VALUE = character(1),
         USE.NAMES = FALSE,
         FUN = function(id) {
-          ret <- headshot_map$espn_headshot[headshot_map$savant_id == id]
-          if(length(ret) == 0 || is.na(ret)) ret <- na_headshot(na_headshot_to_logo)
-          ret
+          url <- paste0("https://midfield.mlbstatic.com/v1/people/", id, "/mlb/436?circle=true")
+          if (isFALSE(check_url(url))) {
+            url <- paste0("https://midfield.mlbstatic.com/v1/people/", id, "/spots/436")
+          }
+
+          return(url)
+        }
+      )
+      gt::web_image(image_urls, height = height)
+    }
+  )
+
+}
+
+
+#' @rdname gt_mlb_headshots
+#' @export
+gt_fmt_mlb_dot_headshot <- function(gt_object,
+                                columns,
+                                height = 30,
+                                locations = NULL) {
+
+  stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
+
+
+  rlang::check_installed("gt (>= 0.8.0)", "to render images in gt tables.")
+
+  if(is.null(locations)){
+    locations <- gt::cells_body({{ columns }})
+  }
+
+  gt::text_transform(
+    data = gt_object,
+    locations = locations,
+    fn = function(mlb_id){
+      image_urls <- vapply(
+        mlb_id,
+        FUN.VALUE = character(1),
+        USE.NAMES = FALSE,
+        FUN = function(id) {
+          paste0("https://midfield.mlbstatic.com/v1/people/", id, "/spots/436")
+        }
+      )
+      gt::web_image(image_urls, height = height)
+    }
+  )
+
+}
+
+
+
+
+#' @rdname gt_mlb_headshots
+#' @export
+gt_fmt_milb_dot_headshot <- function(gt_object,
+                                     columns,
+                                     height = 30,
+                                     locations = NULL) {
+
+  stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
+
+
+  rlang::check_installed("gt (>= 0.8.0)", "to render images in gt tables.")
+
+  if(is.null(locations)){
+    locations <- gt::cells_body({{ columns }})
+  }
+
+  gt::text_transform(
+    data = gt_object,
+    locations = locations,
+    fn = function(mlb_id){
+      image_urls <- vapply(
+        mlb_id,
+        FUN.VALUE = character(1),
+        USE.NAMES = FALSE,
+        FUN = function(id) {
+          paste0("https://midfield.mlbstatic.com/v1/people/", id, "/milb/436?circle=true")
         }
       )
       gt::web_image(image_urls, height = height)
@@ -198,23 +444,25 @@ gt_fmt_mlb_headshot <- function(gt_object,
 #' Replace Team Abbreviations/Player IDs With Images In Column Labels
 #'
 #' @description
-#' `gt_column_example` takes in a value of a team abbreviation or player id and
+#' `gt_mlb_column_labels` takes in a value of a team abbreviation or player id and
 #' converts the designated column to the corresponding image.
 #'
 #' @param value What team abbreviation/player id should be replaced with an image?
 #' @param type What type of image is replacing the value?
 #' @param height The absolute height (px) of the image
 #' @param na_headshot_to_logo should NA/non player id matches return the MLB logo instead
-#'   of a grayed out blank headshot? Defaults to `TRUE`
+#'   of a grayed out blank headshot? Ignored unless `value` is equal to `"headshot"`. Defaults to `TRUE`
 #'
 #' @return HTML tag for image
 #' @export
 #' @examples
+#' \donttest{
 #' library(gt)
 #' library(mlbplotR)
 #'
 #' df <- data.frame(BAL = 1,
 #'                  TEX = 1,
+#'                  LAD = 1,
 #'                  "Mike_Trout" = 1,
 #'                  "Shohei_Ohtani" = 1
 #'                  )
@@ -223,28 +471,39 @@ gt_fmt_mlb_headshot <- function(gt_object,
 #'   gt::gt() %>%
 #'   gt::cols_label(BAL = gt_mlb_column_labels("BAL", "mlb_logo"),
 #'                  TEX = gt_mlb_column_labels("TEX", "scoreboard_logo"),
-#'                  "Mike_Trout" = gt_mlb_column_labels(545361, "headshot"),
+#'                  LAD = gt_mlb_column_labels("LAD", "dot_logo"),
+#'                  "Mike_Trout" = gt_mlb_column_labels(545361, "dot_headshot"),
 #'                  "Shohei_Ohtani" = gt_mlb_column_labels(660271, "headshot"))
+#' }
 #' @export
 gt_mlb_column_labels <- function(value,
                                  type = c("mlb_logo", "scoreboard_logo",
-                                          "headshot"),
+                                          "dot_logo",  "headshot", "dot_headshot"),
                                  height = 30,
                                  na_headshot_to_logo = TRUE) {
 
   if (type == "headshot") {
 
-    headshot_map <- load_headshots()
+    hs_url <- paste0("https://midfield.mlbstatic.com/v1/people/", value, "/mlb/436?circle=true")
 
-    hs_url <- headshot_map$espn_headshot[headshot_map$savant_id == as.numeric(value)]
-    if(length(hs_url) == 0 || is.na(hs_url)) hs_url <- na_headshot(na_headshot_to_logo)
-
+    if (isFALSE(check_url(hs_url))) {
+      hs_url <- paste0("https://midfield.mlbstatic.com/v1/people/", value, "/spots/436")
+    }
 
     html_content <- paste0("<img src=\"", hs_url, "\" style=\"height:", height, "px;\">")
 
     html_content <- gt::html(html_content)
     html_content
 
+  } else if (type == "dot_headshot") {
+
+    hs_url <- paste0("https://midfield.mlbstatic.com/v1/people/", value, "/spots/436")
+
+
+    html_content <- paste0("<img src=\"", hs_url, "\" style=\"height:", height, "px;\">")
+
+    html_content <- gt::html(html_content)
+    html_content
   } else {
 
     team_abbr <- clean_team_abbrs(as.character(value), keep_non_matches = FALSE)
@@ -283,6 +542,7 @@ gt_mlb_column_labels <- function(value,
 #' @param color The color for the top text.
 #' @return An object of class `gt_tbl`.
 #' @examples
+#' \donttest{
 #' library(gt)
 #' library(mlbplotR)
 #'
@@ -293,6 +553,7 @@ gt_mlb_column_labels <- function(value,
 #'   gt_merge_stack_team_color(col1 = "team_location",
 #'                             col2 = "team_mascot",
 #'                             team_col = "team_abbr")
+#' }
 #' @export
 gt_merge_stack_team_color <- function (gt_object,
                                        col1,